Add v1ScsPluginName to convertResult object
Remove circular dependency problems
